+++ AWSClientRuntime/Tests/Http/AWSUserAgentMetadataTests.swift
@@ -6,6 +6,7 @@
 //
 
 import XCTest
+import ClientRuntime
 @testable import AWSClientRuntime
 
 class AWSUseragentMetadataTests: XCTestCase {
@@ -80,5 +81,36 @@ class AWSUseragentMetadataTests: XCTestCase {
                                        frameworkMetadata: frameworkMetadataWithExtras)
 
         XCTAssertEqual("aws-sdk-swift/0.0.1 api/meow/1.1 os/iOS/13.1 lang/swift/5.0 md/test1/1.2.3 exec-env/e123 lib/aws-amplify/2.0.1 md/f1/c1", sut.xAmzUserAgent)
+    }
+
+    func testUserAgent() {
+        let ua = AWSUserAgentMetadata.fromEnv(apiMetadata: APIMetadata(serviceId: "Test Service", version: "1.2.3"))
+        XCTAssert(ua.userAgent == "aws-sdk-swift/1.2.3")
+    }
+    
+    func testXAmzUserAgent() {
+        let currentOS = ClientRuntime.currentOS
+        let apiMeta = APIMetadata(serviceId: "Test Service", version: "1.2.3")
+        let sdkMeta = SDKMetadata(name: "swift", version: apiMeta.version)
+        let osMeta = OSMetadata(family: currentOS, version: "11.4")
+        let langMeta = LanguageMetadata(version: "9.9")
+        let awsUserAgent = AWSUserAgentMetadata(sdkMetadata: sdkMeta,
+                                      apiMetadata: apiMeta,
+                                      osMetadata: osMeta,
+                                      languageMetadata: langMeta)
+        var expected: String = ""
+        switch currentOS {
+        case .linux:
+            expected = "aws-sdk-swift/1.2.3 api/test-service/1.2.3 os/linux/11.4 lang/swift/9.9"
+        case .macOS:
+            expected = "aws-sdk-swift/1.2.3 api/test-service/1.2.3 os/macOS/11.4 lang/swift/9.9"
+        case .iOS:
+            expected = "aws-sdk-swift/1.2.3 api/test-service/1.2.3 os/iOS/11.4 lang/swift/9.9"
+        case .unknown:
+            expected = "aws-sdk-swift/1.2.3 api/test-service/1.2.3 os/unknown/11.4 lang/swift/9.9"
+        default:
+            XCTFail("Unexpected OS")
+        }
+        XCTAssert(awsUserAgent.xAmzUserAgent == expected, awsUserAgent.xAmzUserAgent)
     }
 }
